Wsparcie eventów, dodanie zmiennej {POS}, usprawnienie kodu bugujących się bloków

+++ src/main/java/net/dzikoysk/funnyguilds/command/ExcConfirm.java
@@ -1,5 +1,9 @@
 package net.dzikoysk.funnyguilds.command;
 
+import org.bukkit.Bukkit;
+import org.bukkit.command.CommandSender;
+import org.bukkit.entity.Player;
+
 import net.dzikoysk.funnyguilds.basic.Guild;
 import net.dzikoysk.funnyguilds.basic.User;
 import net.dzikoysk.funnyguilds.basic.util.GuildUtils;
@@ -9,9 +13,9 @@ import net.dzikoysk.funnyguilds.data.Settings;
 import net.dzikoysk.funnyguilds.data.configs.MessagesConfig;
 import net.dzikoysk.funnyguilds.data.util.ConfirmationList;
 import net.dzikoysk.funnyguilds.data.util.MessageTranslator;
-import org.bukkit.Bukkit;
-import org.bukkit.command.CommandSender;
-import org.bukkit.entity.Player;
+import net.dzikoysk.funnyguilds.event.FunnyEvent.EventCause;
+import net.dzikoysk.funnyguilds.event.SimpleEventHandler;
+import net.dzikoysk.funnyguilds.event.guild.GuildDeleteEvent;
 
 public class ExcConfirm implements Executor {
 
@@ -44,6 +48,11 @@ public class ExcConfirm implements Executor {
         ConfirmationList.remove(user.getUUID());
 
         Guild guild = user.getGuild();
+        if (!SimpleEventHandler.handle(new GuildDeleteEvent(EventCause.USER, user, guild))) {
+            return;
+        }
+        
+        
         GuildUtils.deleteGuild(user.getGuild());
 
         MessageTranslator translator = new MessageTranslator()
